PR: Succès & Trophées

+++ update-staging.sh
@@ -1,7 +1,8 @@
 #\bin\bash
+source .env
 
-php artisan down
 php artisan service:locked
+php artisan down
 
 git reset --hard
 git pull origin master
@@ -9,7 +10,19 @@ git pull origin master
 composer install --prefer-dist --no-interaction
 npm install
 
-php artisan migrate:fresh --seed
+if [ "$(git diff --name-only HEAD HEAD~1 -- database/migrations)" != "" ]
+then
+    mysql -u votre_nom_utilisateur -p votre_mot_de_passe -e "
+
+    # Suppression de la base de données
+    DROP DATABASE IF EXISTS ${DB_RAILWAY_DATABASE};
+
+    # Création de la base de données
+    CREATE DATABASE ${DB_RAILWAY_DATABASE};
+    "
+    php artisan migrate:fresh --seed
+fi
+
 php artisan release:update
 php artisan cache:clear
 php artisan config:clear
@@ -20,5 +33,5 @@ php artisan webpush:vapid
 php artisan horizon:terminate
 chmod -R 777 storage bootstrap/cache
 
-php artisan service:unlocked
 php artisan up
+php artisan service:unlocked
